+++ tasks/jshint.js
