+++ tasks/support/.jshintrc
